video-downloader-bot/requirements.txt
pyTelegramBotAPI>=4.14.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
